+++ discord_twitter_webhooks/main.py
@@ -299,4 +299,4 @@ def sched_func() -> None:
 
 
 if __name__ == "__main__":
-    uvicorn.run(app, host="0.0.0.0", port=8000)
+    uvicorn.run(app, port=8000)
